requirements.txt
beautifulsoup4==4.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
future==0.18.2
idna==2.8
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pycparser==2.20
pytesseract==0.3.7
python-dateutil==2.8.1
python-telegram-bot==12.3.0
pytz==2021.1
requests==2.22.0
six==1.15.0
soupsieve==2.2.1
tornado==6.1
urllib3==1.25.11
regex==2021.4.4
